requirements.txt
numpy
pandas
matplotlib
scikit-learn
gradio
paho-mqtt
python-dotenv